split up ui package
