+++ app/src/main/java/com/example/fitnesstracker/ui/progress/ProgressionFragment.java
@@ -1,4 +1,4 @@
-package com.example.fitnesstracker.ui;
+package com.example.fitnesstracker.ui.progress;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,13 +8,13 @@ import androidx.fragment.app.Fragment;
 
 import com.example.fitnesstracker.R;
 
-public class NutritionFragment extends Fragment {
+public class ProgressionFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Layout für das Fragment laden
-        View view = inflater.inflate(R.layout.fragment_nutrition, container, false);
+        View view = inflater.inflate(R.layout.fragment_progression, container, false);
 
         return view;
     }
